file streaming now getting correct file size
We had an issue with csv downloads failing due to an incorrect content
length. This might have been previously hidden by some fastcgi
translation/content handling, but after switching to waitress/wsgi
directly we see waitress erroring on the incorrect content length set by
the app.

The root cause is that the length of a byteIO object is not actually the
length of the content. byteIO dynamically allocates memory pages, which
may be larger than the content in those pages. If we instead seek to the
last byte in the content we can very efficiently get the actual content
length.